Don't use a .gitnore unless you actually need one.
It's important to make use of every file in a repo. If there are no files that should be ignored, don't use a .gitnore file. This will also help with faster run time speeds and less clutter, saving you time and headache when trying to find the file you're looking for.